Split Profile into Profile/ProfilePage
